Update .gitattributes
Guess the Number Game

This Python script is a simple "Guess the Number" game where the player tries to guess a secret number within a limited number of attempts. The game's rules are straightforward:

The computer selects a secret number.
The player has a maximum of three attempts to guess the secret number.
If the player correctly guesses the secret number within the given attempts, they win the game.
If the player exhausts all three attempts without guessing the correct number, they lose the game.
How to Play:

Clone this repository to your local machine or download the code.
Run the script in a Python environment.
Follow the prompts to enter your guesses.
You'll receive feedback on whether your guess is correct or not.
Feel free to modify and enhance the game to suit your preferences or add more features. Have fun playing and learning with this basic Python guessing game!